Consolidated some test files and made more consistent
