--- pkg/modelregistry/jsonvalues/testdata/sample-device1.json
@@ -1,8 +1,0 @@
-{
-  "cont1a": {
-    "cont2a": {
-      "leaf2a": 12,
-      "leaf2b": 1.21
-    }
-  }
-}
